+++ project/osc-front/src/views/home/index.vue
@@ -1,21 +1,21 @@
 <template>
-    <div class="main-container">
+    <div class="home-main-container">
         <el-row :gutter="20">
             <!-- 左侧主区块，占据75%，用于仪表盘展示 -->
             <el-col :span="18" :offset="0">
-                <div class="left-content">
+                <div class="home-left-content">
 
                 </div>
             </el-col>
 
             <!-- 右侧侧边区块，占据25%，用于快捷功能和最近访问 -->
             <el-col :span="6" :offset="0">
-                <div class="right-content">
+                <div class="home-right-content">
                     <!-- 分为上下两个，各占50% -->
-                    <div class="right-top">
+                    <div class="home-right-top">
 
                     </div>
-                    <div class="right-bottom">
+                    <div class="home-right-bottom">
 
                     </div>
                 </div>
@@ -29,23 +29,23 @@
 </script>
 
 <style scoped>
-.main-container {
-    min-height: 100vh;
+.home-main-container {
+    min-height: 100%;
     display: flex;
     flex-direction: column;
 }
 
-.main-container, .el-row {
+.home-main-container .el-row {
     flex: 1;
     display: flex;
 }
 
-.main-container, .el-col {
+.home-main-container .el-col {
     display: flex;
     flex-direction: column;
 }
 
-.left-content {
+.home-left-content {
     flex: 1;
     background-color: #f5f7fa;
     border-radius: 5px;
@@ -53,23 +53,23 @@
     min-height: 300px;
 }
 
-.right-content {
+.home-right-content {
     flex: 1;
     display: flex;
     flex-direction: column;
 }
 
-.right-top, .right-bottom {
+.home-right-top, .home-right-bottom {
     background-color: #f5f7fa;
     border-radius: 5px;
     padding: 20px;
 }
 
-.right-top {
+.home-right-top {
     flex: 1;
 }
 
-.right-bottom {
+.home-right-bottom {
     flex: 1;
     margin-top: 20px;
 }
